Pull out saving settings into a separate util
